App integrates two or more Google services. Google service integrations can be a part of Google Play Services or Firebase.
Each service imported in the build.gradle is used in the app.

If Location is used, the app customizes the user’s experience by using the device's location.

If Admob is used, the app displays test ads. If Admob was not used, student meets specifications.

If Analytics is used, the app creates only one analytics instance. If Analytics was not used, student meets specifications.

If Maps is used, the map provides relevant information to the user. If Maps was not used, student meets specifications.

If Identity is used, the user’s identity influences some portion of the app. If Identity was not used, student meets specifications.

DONE.